requirements.txt
# AI Overkill - Python Dependencies
#
# Python 3.10+ required (matches pyproject.toml)
#
# Installation:
#   python3 -m venv .venv
#   source .venv/bin/activate
#   pip install -r requirements.txt
#
# Most hooks and scripts use only the standard library:
# - json, os, sys, re (core)
# - sqlite3 (learning database)
# - subprocess, shlex (command execution)
# - pathlib (path handling)
# - dataclasses (structured data)
# - datetime (timestamps)
# - hashlib (signatures)
# - ast (code analysis)
# - contextlib (resource management)
# - shutil (file operations)
# - uuid (unique identifiers)

# Required for evals harness (agent/skill grading)
PyYAML>=6.0

# Required for WordPress integration scripts (optional)
requests>=2.28
python-dotenv>=1.0

# Required for Google Search Indexing API (optional)
google-auth>=2.22
